requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_full_version >= '3.6.0'
altgraph==0.17
apscheduler==3.6.3
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.9.3
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1; python_full_version >= '3.6.0'
dawg-python==0.7.2
docopt==0.6.2
feedparser==6.0.8
filelock==3.0.12
huggingface-hub==0.0.12; python_full_version >= '3.6.0'
idna==3.2; python_version >= '3'
iniconfig==1.1.1
joblib==1.0.1; python_full_version >= '3.6.0'
loguru==0.5.3
mock==4.0.3
multidict==5.1.0; python_full_version >= '3.6.0'
navec==0.10.0
numpy==1.21.1; python_version >= '3.7'
packaging==21.0; python_full_version >= '3.6.0'
pandas==1.3.1
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.8.2; python_full_version >= '3.6.1'
pyinstaller-hooks-contrib==2021.2
pyinstaller==4.5
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2==0.9.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytelegrambotapi==3.8.2
pytest-mock==3.6.1
pytest==6.2.4
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-telegram-bot==13.7
pytz==2021.1
pyyaml==5.4.1
razdel==0.5.0
regex==2021.8.3
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.7.1; python_version < '3.10' and python_version >= '3.7'
sentencepiece==0.1.96
sgmllib3k==1.0.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slovnet==0.5.0
soupsieve==2.2.1; python_version >= '3'
ta==0.7.0
threadpoolctl==2.2.0; python_full_version >= '3.6.0'
tinvest==3.0.5
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
tornado==6.1; python_version >= '3.5'
tqdm==4.62.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.9.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4.0'
xgboost==1.4.2
yarl==1.6.3; python_full_version >= '3.6.0'
